Add helper script for running flake8 when developing
This adds the command `make flake8` to check the local files in this
project for lintable problems.  I like this because it means all
developers will be using the same procedure for flake8, and this will
also be the same version of that software employed by CI.

I also tried to implement this in a way that opens the door for easily
adding `make` commands that are, essentially, shell commands that
should be run in the devops environment.