Adds pshtt pip dependency
Installing 0.2.3, which is the latest release, and unfortunately the
ONLY release available on pip, because the pshtt maintainers aren't
doing a good job with package publishing desiderata.

We're installing the pshtt dependency via python3, inside the webapp
virtualenv. There's a version conflict with elasticsearch, so we circle
back and update elasticsearch after the fact.