Address deficiency docker_image change detection
I've noticed that when you call docker_image with a name and a tag that doesnt
change BUT the underlying dockerfile changes, the module is not smart enough to
rebuild the image. So I've tweaked the tagging logic a little to take a piece of
checksum from the requirements + the dockerfile.