Consolidates pshtt across dev and ci scenarios
Plays with the molecule testinfra config parameters so that both the `dev` and
`ci` scenarios can utilize the same `test_pshtt.py` file. I also slightly
tweaked the test to output the csv to a temporary location instead of the
current directory - to avoid having ghost `results.csv` hanging around.