Move relevant files to top-level within `landing_page_checker` app
This commit restructures the folders within the new
`landing_page_checker` app so that it behaves more like a subset of
functionality within SecureDrop.org devoted to checking landing pages
and less like the stand-alone Django app it began its existence as.

Thus most of the files here now live in the top-level folder within
the app.  I've also preserved `landing_page` as a subfolder, as this
organization seems useful.