Fixed tests and code to reflect new app structure
Re-do module imports so that the `landing_page_checker` code now
correctly knows the locations of all its own components, that's to say
it now imports everything from the `landing_page_checker` module, not
the `directory` module.  This should cause all tests to pass _except_
the view tests, which I removed since they were testing specific
template responses, and these templates are no longer going to be in
use.

It's actually possible the views themselves will not be in use, but
I haven't really determined that yet.