Enable instance-scanning when DirectoryForm submitted
Use the landing page checker code to actually scan the instance and
record the results.

Note that this commit removes our dummy function to test whether an
instance is valid.  We don't want this around anymore since basically
want to record all (or at least most) scan results and put them in the
directory regardless of whether it passes the test.