Consider `vcrpy` part of the non-dev requirements
The test suite is a useful development tool, but it is also specifying
the actual capabilities of the software.  So it's not clear to me
whether or not testing-only requirements such as VCR.py should be in
the development requirements.  I'm moving it out and into the main
requirements file until I can convince myself otherwise.