Initial ScanForm
With this form, we can now "just" scan an instance's landing page
without saving anything to our database.  This scans the site and
reports the results with the grade.

There are no styles, and there are some places where we could tighten
the code up a little, but this is just a start.