Form at 'scan/' properly renders errors.
If form is invalid, the form is rendered with errors instead of the previous 'not cool'
It's a bit hard to tell from the diff, but if the method is POST, the form is set to the current form, and if it's not valid, that form will be rendered with errors. A new form is now explicitly shown only on a GET request.