Require 2FA verification to view the securedrop admin dashboard
This commit adds some somewhat confusing logic to MyAccountAdapter,
which I've attempted to explain via comments, but the overall gist is
that our desired workflow is for the user to register, immediately
configure 2FA, then (and only then) be permitted to manage their
Securedrops.

There is a little bit of conflict between this desired flow and how
django_otp, allauth, and django_allauth_2fa all play together here,
and I think I've minimized the intrusiveness into the inner workings
of all three our own code has to be to ensure that this works for
newly registered users as well as returning users.