Updates Makefile target for pip reqs
Ensures that we build Python3 requirements inside a Python3 environment
(for consistency with the intended prod webserver). The dev/CI reqs are
Python2, so those can be compiled via `pip-compile` locally (not in a
container).

Using `--no-header` because the default headers are misleading: the
correct workflow is to read the docs and rely on Makefile targets to Do
The Right Thing™.

Also removed the manual include from the dev-reqs file: pip-compile
supports merging multiple requirements files out of the box, so let's
let pip-compile to the heavy-lifting, and explicitly combine them by
passing multiple files to pip-compile via the Makefile target.

Purges old bash script for updating pip dependencies, so as not to
confuse.