Fix test of non-live site by mocking requests.get
Previously this test was broken: it would work if no cassette was
saved and it would record a new cassette, but it would not record
the non-existent websites to the cassette (presumably because vcrypy
does not record DNS failures). On the second run it would try to
look up the non-existent site in the cassette and, not finding it,
would throw an exception. We work around this by mocking requests.get
to throw a ConnectionError immediately when receiving our special
non-existent site URL and not using vcr in that case at all.