Pins Molecule at v2.12 for CI tests
Our use of an env var for a fullpath to the pshtt binary is causing the
overly aggressive Molecule validator to fail, since it expects env var
values to be strictly alphanumeric. That's practically insane. Will open
an issue over there and see if we can't get that cleaned up. Selectively
downgraded from 2.16 to identify 2.12 as the last working version for
us.