Fix ScanResult to DirectoryEntry association on save
At some point this association was changed to a foreign key from
whatever it was before that necessitated saving the association in
this way.  This should restore the correct behavior and ensure any
future code that needs it works, such as the tests for the code in
this branch.

See commit a0912c6391d33996439d89d210cf986eb653c0c6