Add strings file for holding warning message strings
These are (mostly) placeholders for slightly more informative
messages, but I wanted them to be somewhat flexible and be able to
reference attributes on (or derived from) the DirectoryEntry or
ScanResult objects.  Thus I think it's better to have them be stored
as python values rather than as part of the page templates.