Display warnings on DirectoryEntry pages
Here I'm attempting to match the mock-ups that were made for this
feature.  My other goal was not to touch the style code too much, for
fear of breaking something somewhere else.  I think it's pretty
close.  One thing that remains to be done is update the link to the
landing page itself to be the color of the warnings.