Add ChoiceArrayField
I'm adding this field to get around what I belive is a UI deficiency
within wagtail, namely that it's not easy to select multiple choices
if it's not some kind of many-to-many model relationship.  Sometimes
you just want a choice field where the person can pick multiple
options, and those options are stored in a postgresql array.  The
default UI for that is a single line, comma separated entry, which I
think this ChoiceArrayField is superior to.

Added in preparation for choosing warnings to ignore on directory
entry pages.