scanner: Ensure that `forces_https` is either true or false
Here, `pshtt` sometimes returns `None` in the case where the landing
page domain's non-HTTPS responses are not redirects.  We consider this
a failure of our forces-http test, so we want to ensure those cases
are converted to `False` before being saved.

Test added with a VCR recording of a domain that exhibited the quirk
of having `None` in `pshtt_results['Strictly Forces HTTPS']`.

Refs #532