scanner: Permit scanner to follow redirects
Changes the scanner to follow redirects from the given landing page
url to a final destination.  The final destination is saved to the
scan result and this page (and its domain) are inspected for
compliance.

This changes the order in which the scan operations occur.
Previously, we would scan the domain and then make a normal request to
the page for its content and headers.  Now we must make the request
first, follow any redirects, then scrape and scan the target domain.

I've taken this reordering as an opportunity to slightly change how
we're construcing our `ScanResult` object.  Namely, I've split the
processing of the three pieces of data that compose it (the http
response, the page content, and the domain information) into three
separate functions that take the scan data and return a dictionary of
`ScanResult` attributes that map to boolean values indicating
pass/fail.  The `perform_scan` function merges these dicts into the
final `ScanResult`.

The benefits of this are smaller and more understandable data
processing code, and a shared back-end to both `scan` and `bulk_scan`,
which prior to this change both implemented their own mostly-similar
fetch-and-process code.  Now `perform_scan` does that for everyone,
and it does so by passing off those three pieces of data to their own
processing functions.  I think this design is cleaner and more
understandable, but I'm open to comments and criticism or other
feedback about what this change does.