Rename attributes on Asset object
For human-readable summaries of 3rd party assets, it's helpful to have
slightly more information than we were gathering prior to this.  The
`initiator` (verbiage copied from Chrome's web inspector) attribute is
the name of the file or resource that initiated the request (or
potential request) for the given asset.  This means we can more easily
track down exactly where a given asset is found on the landing page.