Use session storage to remember tor warning dismissal
Session storage differs from session cookies only in that, I believe,
opening the page in a new tab or window will start a new session.  So
in theory this will display the warning more than using a cookie (and
thus be more annoying).  But I prefer the session storage approach for
its cleaner API and that this seems like a use that session storage
was designed for.

Refs #101