Ignore assets of type script-resource or script-embed
These are assets we've obtained by examining Javascript code for
patterns that look like URLs, but this approach generates a lot of
false positives in the form of non-URLs (strings that look like URLs
but are not) and URLs that are not used to make requests (which is the
thing we're actually scanning for).  So let's ignore these for now,
log them, and re-assess if we come up with a better solution for
parsing or executing landing page Javascript.