Allow arbitrary domains to be permitted by `parse_asset`
This change generalizes one that was formerly specific: ignoring
assets on specific domains when checking for cross-domain assets.  We
used to examine all well-formed assets on the page and skip over any
that were on the same domain as the landing page.  Now we skip over
any that are on a list supplied by the caller which is populated by
default with the landing page domain, plus any others that belong to
the DirectoryEntry object.

It's possible we want to further generalize this by having
`perform_scan` accept some kind of `ScanConfig` object that has
everything you need to conduct a scan.  I dunno!  For now I actually
like how we can do more with simpler code.

Of note: permitted domains are top level domains only.  You can't
permit a subdomain.  Permitting 'x.y.z.com' will permit everything on
'z.com', because, for security and privacy purposes, subdomains are
the same as top level domains.