scanner: Remove early return if cross domain redirect detected
Previously, if we found a cross-domain redirect, we'd avoid expending
the extra resources continuing to scan the domain to check its assets
and if HTTPS was configured securely.  But on further consideration,
doing this is not all that expensive and it gives us a more complete
look at what's going on with the landing page.  Also, certain other
components of our system involve the `live` attribute, which is only
updated by `pshtt`, so we should always try to get this to be set
correctly, even if the target domain is different from the landing
page domain.