scanner: Ensure that asset scanning is done using the target url
`url` is the url given as an argument to the `perform_scan` function.

`page.url` is the url of the `Response` object in which is stored the
url of the final destination of our original request, taking into
account all redirections.

We want to use `page.url` here, since it's possible there have been
cross-domain redirects, so the url used for asset scanning might be
fairly different from the one initially fetched.