Make category a required field for blog pages
Rationale: The UI has a defined place to display the name of the
category with a link to the category page and it breaks (with a server
error) when trying to render a blog page without a category.  So if we
want to enable free-floating blog posts without categories, the
template will have to be revised.  Until then, a category should be
required.