directory: Remove the "warnings" querystring flag
This change will make it so that anyone visiting the page will see the
warnings, and a flag in the URL is not required.

The "show_warnings" flag remains in the template because there is
still at least one case were we do not want to try to show warnings:
when there are no scan results for that directory entry.  So I think
this small bit of logic is fine to leave in.