Resolve psql dev errors
Honestly, I dont understand how this was ever working before. I suspect
that at somepoint the underlying image was altered in a way where the
default `postgres` user is now stripped out. If you don't specify a `-U`
username the container will assume local username. In this case I was
getting errors about a role for `postgres` not existing.