Allow warnings to appear correctly in wagtail previews
This change fixes a crash that happened when previewing directory
entries.  The wagtail preview feature uses `FakeQuerySet` objects
that, apparently as of at least version 1.12, don't implement the
`latest` function.

So since `latest` is actually syntactic sugar for some sorting logic,
I converted our problematic function to use the lower level queryset
functions instead, which are compatible with previewing.  This change
should produce identical results to the code it's replacing.