Dev env working in docker-compose
Wires up scripts to allow developers to fire up a dev environment
quickly using `docker-compose`. Removes previous molecule dev
environment.

Note here that we had to do this awkward UID injection into
`docker-compose` to address differences in macOS vs. Linux.
`docker-compose` has a concept of using env var substitution in the
docker-compose.yaml file BUT macOS (as far as I can determine) doesn't
have a good env-var exposed to get the UID and even if it does, it needs
the exact same one as used on Linux for it to work. Sooo instead we have
to do this hacky injection process on first run. I'd love to figure out
an alternative here. Hoping open-sourcing it will bring in a new idea.