Fix double-firing signal from wagtail factory creation
I'm not totally sure why this was happening, but there was a test of
"on blog page save" signal that was firing twice when that model was
being created, in the test, from a factory.  I've changed it so we're
not creating it that way anymore and, lo and behold, the test passes
as it's supposed to.  I'm going to chalk this up to some kind of
weirdness in the factory library itself and move on.