Bumps nodejs version 6.11.0 -> 10.15.1
Updates applied to multiple files. Converted the node version in the
prod django container (which is a multi-stage build) to a required build
arg, by removing the default. Doing so forces us to pass in the desired
version at build time. Did this because I'd initially overlooked the
hardcoded default.