Runs `npm audit` in CI
Checks the pinned node dependencies, as specified in
`package-lock.json`, for vulnerabilities. Running as part
of the "safety" workflow in CI, so deps are validated before
the tests are run (for faster feedback).

Note that `npm audit` currently outputs a list of known-bad deps, so
this change will cause CI to fail. We'll have to update the dependencies
in a PR in order to get CI passing, at which point we can resolve any
breaking changes to assets, and merge.