Fix use of mock objects in tests for python 3.5
Python 3.5 helps fix an (apparently) common problem of mistyping
method names on mock objects.  Since a mock object is designed to
respond to any methods you give it, typoing a name can cause some
silent failures.  Specifically, methods we expect to raise an
exception will not do so.

So, `mock` doesn't have a method named `assert_called_once`.  In
python 3.4 writing `mock_object.assert_called_once()` not do anything,
but in 3.5 it raises an exception saying you probably intended to call
a different method.  I've replaced all uses `assert_called_once` with

```
assertEqual(mock_object.call_count, 1)
```

Which asserts the same thing but is unlikely to fail silently due to
typo, since one is less likely to accidentally set a typo'd variable
to 1.

Doing this revealed that some of the systems being tested in this way
were not being tested correctly, as the tests calling
`assert_called_once` were never failing.  I've made a few changes here
and there to fix that.

In one case, the cache purging signal for when settings were changed,
the system itself appeared to not be working correctly.  I think
`sender=<some base class>` doesn't work for settings the same way it
does for models?  Either way, changing the signal to a "universal"
signal with an `issubclass` check within was the only way I could get
those tests to pass.