Add monkey-patch workaround for nested block crash in draftail
There's a bug in the current release of wagtail that makes its new
rich text editor not backwards-compatible with some HTML that could be
generated by the old rich text editor.  This code attempts to work
around the issue to avoid the crash by monkeypatching the HTML parser
within the wagtail rich text system.

Ideally this code can be removed when the bug itself is fixed.  There
is a test for the "original" wagtail behavior that will fail when the
parsing no longer raises an error.  At this point we can remove this
patch.

See discussion here: https://github.com/wagtail/wagtail/issues/4602