Ensure that frontend caches don't cache form pages
CSRF tokens are distributed on form pages in the response on a per
user basis. A frontend cache like cloudflare can interfere with
this process by displaying the same token to multiple users due to
caching the response. The form then throws an error when the token
doesn't validate.

This change specifies that form pages are "private"--only to be
cached by the end-user's browser, not intermediaries. With this
change, cloudflare should not store form page responses at all,
resolving the CSRF issue.

See:

- https://support.cloudflare.com/hc/en-us/articles/115003206852-Enabling-Origin-Cache-Control-with-Cloudflare-Page-Rules
- https://docs.djangoproject.com/en/2.2/ref/csrf/