Create separate requirements file for pip tools
It's possible this is more overhead than we want, but it also seemed
to me unnecessary to install the entire dev-requirements.txt onto this
container when all we really want to use for these
dependency-compiling make commands is `pip-tools` itself.  So I made a
separate file for it.  If this is too much, let me know.