Replace VCS-based requirements with github-hosted zip archives
See https://github.com/jazzband/pip-tools/pull/807 for more
information about permitting URLs as packages.  We want this because
--generate-hashes works with URL-based requirements but not with
Git-repo-based requirements.  Subtle different, but it works.  As such
I think it's okay to remove the separate github URL requirements file.

One other thing I do here is reconcile the pshtt/ssylze/cryptography
dependency-hell situation by pinning the version of sslyze to the
latest one that supports (a) the most recent, secure version of
cryptography, and (b) python 3.5.  Some of the versions of these
packages were not release the last time this situation was examined.