Make `ReadOnlyPanel` more future-proof
Something changed internally in the Wagtail 2.5 version of
`EditHandler` and it was causing the site to not start.  I've updated
this class to, I hope, be slightly more future proof by calling
`__init__()` on the superclass and making the child class's
`__init__()` method signature take any number of arguments and pass
them through to `EditHandler`.

Will this prevent future breakages?  I have no idea.  But for certain
types of changes to `EditHandler` I think it will help.