Add Inline PDF block for blog pages
This block for explicitly rendering a PDF inline attempts to
compromise between

1. Browsers that don't support inline PDFs (mobile safari, chrome for
android).  See
https://en.wikipedia.org/wiki/Comparison_of_web_browsers#Image_format_support

2. Browsers with javascript disabled (Tor Browser in Safest mode)

In both cases, I think we want to provide a fallback in the form of a
link to download the PDF directly.  To handle the browser support
detection, I've added a JS library called `pdfobject` which handles that
and also allows us to make all inline-PDF rendering contingent on JS
being enabled and provide a single fallback for that in the form of a
`<noscript>` tag.

It is theoretically possible to render PDFs inline without JS using
`<iframe>` or `<object>` or `<embed>` but it's hard to have these
methods gracefully fail if the browser does not support PDFs.
Furthermore, since some browser PDF rendering is done _using_
JS (e.g. pdf.js in Firefox/Tor Browser), there are further
complications for directly embedding the PDF without using JS first,
as the result can be a blank box or PDF UI frame that contains
nothing.  If we use JS to load the PDF itself, then we can
short-circuit the whole process if JS is not supported, and then we
control exactly what is displayed as a fallback.