webpack: Remove references to modernizr and modernizr-loader
This tool is not being used (there is no .modernizrrc file in our
repo) and does not need to be installed as part of package.json.  I'm
removing it because right now it has a security vulnerability in one
of its dependencies:

https://snyk.io/vuln/SNYK-JS-MINIMIST-559764

but in general we shouldn't install software we no longer need (or
never needed).