middleware: Add the Onion-Location header from settings
This adds a new piece of middleware for the single purpose of adding
the Onion-Location header from the site's footer settings, if it
exists.  There's some small logic to ensure that the header value has
a https protocol if what's stored in the settings does not have an
accompanying protocol.