safety: Ignore pillow 6.2.2 vulnerabilities
These are vulnerabilities related to processing of malicious images,
and since we don't expose image uploading to 3rd parties on this site,
somewhat reasonable to ignore.

There is no planned 6.2.3 release of pillow to fix this (see
https://github.com/python-pillow/Pillow/issues/4750) and wagtail 2.7.x
requires this as of today.  Possibly the 2.7 branch of wagail will
relax this requirement, but until then I think it's best to ignore
these.