Improve ignore handling in safety script
What's happening here is I'm updating the operations performed by
`make safety` to:

1. Keep track of which vulnerabilities were ignored and which directly
failed.
2. More easily communicate this information to other scripts.

To this end, I've convered the check into a python script.  This lets
us store the information about what to ignore in python code itself,
potentially with comments.

The script accepts a single command line flag: `--json`.  If given,
the script will print, in JSON format, information about what
vulnerabilities were found, which ones were ignored, and where each of
them occurs in the repo.  This is useful for being processed by
another script, or for a more compact listing of the results of the
check.  Regardless of what vulnerabilities are found, it will exit
with status 0 unless an error occurs during checking.

If `--json` is not provided, the script will run `safety` (the command
line vulnerability checker) on all requirements.txt files.  The safety
command will be instructed to ignore the appropriate vulnerabilities.
This form of the script is intended to produce output readable by
humans, or as part of a CI process.  If vulnerabilities are found, the
script will exit with a nonzero code.