Ensure onion_address setting includes `http://` protocol
The Onion Location Middleware requires this setting to act as a "base"
URL in order to form the correct header value, so we need the
`http://` in this setting.  The validation already tries to make sure
this happens, so we ought to have it in the default value as well.