Remove JSON format
I've decided that the best way to implement this check right now is to
remove as much code as possible to reduce the headache of duplicating
it across all our web repositories.  Since a lot of the logic for
tracking ignored checks vs. failed checks already exists in the
fpf-www-projects safety check tool, I think it's unnecessary to
duplicate it here.

I tried to make this file into a very thin wrapper around `safety`
itself, with the main purpose being to load the appropriate IDs to
ignore from a project-level JSON config file.  The JSON data can be
shared with the fpf-www-projects check and processed to separate
ignores and failures there, rather than here, where this script only
needs to output pass or fail for CI runs on feature branches.