Removes all code directly connected to legacy autocomplete tool
The autocomplete code in our "client" directory has either been
replaced by the wagtail plugin on the admin side, or is related to
the unused "directory management" feature.  This means I think it's
safe to remove.  The motiviation for doing this is the presense of a
security vulnerability in a dependency of react.

It's possible we will want to remove the rest of the directory
management feature at some point, but that I think requires some
discussion.