Put Piwik analytics code into base template directly
To include the settings variables in the template, I've added a
context processor for this specific purpose.

Also, we're ensuring the script follows the CSP by using a nonce and
the django-csp package's `script` tag.  Using this tag requires us to
add `csp` to `INSTALLED_APPS`.

The only change from the code that django-analytical outputs and the
code I've added here is: replace `//` in URLs with `https://`, to
ensure we _always_ make the analytics request over https, even when
the site is being viewed over http, such as in Tor Browser.