Move analytics code to its own JS file
This change removes the analyics code from an inline script tag and
into its own JS file.  My motivation is that now we do not need to
explicitly permit the analyics code in the content security policy.  I
think this is a big security improvement by making the CSP as minimal
as possible.

References:

https://matomo.org/faq/general/faq_20904/