Ensure filter test does not fail based on randomness
This change excludes all languages and countries from the
`other_instance` (the one designed not to show up in the filters under
test).  Without setting languages and countries to 0, they will be
assigned two random ones on creation, which will sometimes coincide
with the countries we're testing in the filter and cause the test to
fail.