Remove "www." prefix from domains before inspecting with pshtt
This is because pshtt expects to be given a base domain for analysis,
and during the process adds its own "www." to domains it is looking
at.  Meaning if we give it "www.example.com" then it will believe this
is the base domain and make requests to "www.www.example.com" which is
incorrect.

The test here heavily mocks, but this is mostly for isolation
purposes, not actually stubbing out behavior.