Attach ad-hoc scan view to the ScanResult admin as create view
Since creating a scan result involves performing an actual scan of a
landing page, it makes sense to me to put this functionality into the
"create view" slot on the pre-existing Scan Result model admin.  The
only customizations I'm really doing here are renaming the "create"
button to something that better describes what happens when you create
a scan.

I also added `landing_page_url` to the fields shown on the scan result
list display, since ad-hoc scans are not related to any directory
entry, so there ought to be some way to distinguish them.