Use correct settings file for prod-like CI configuration
I'm pretty sure we were unintentionally using the `dev` settings for
the "prod-like" CI job and not the production settings.  This corrects
that, and the environment variable duplication is an unfortunate
consequence of how docker handles "build" variables vs. "run"
variables.  If there's another way to do this I'm interested to know
what it is!