Update allauth from 0.45 to 0.47
The previous version was causing `safety` to complain with this
message:

```
-> django-allauth, installed 0.45.0, affected <0.47.0, id 43274

Django-allauth 0.47.0 adds a new setting 'SOCIALACCOUNT_LOGIN_ON_GET'
that controls whether or not the endpoints for initiating a social
login (for example, "/accounts/google/login/") require a POST request
to initiate the handshake. As requiring a POST is more secure, the
default of this new setting is 'False'. This is useful to prevent
redirect attacks.
```

IMO, this part of the site is disabled.  Possibly should be removed at
some point, allowing us to uninstall this package.