Rework DevDjangoDockerfile to use multi-phase builds
Using multi-phase builds here allows us to isolate the image for
building requirements and refer back to it later.  Before this change,
updating the requirements with `make compile pip-dependencies` would
cause docker to reinstall all apt and pip dependencies required for
building our final requirements.txt.  After this change:

1. The builder is set as its own phase, so it can be cached.  We can
extract build artifacts from this phase using `COPY --from=builder`
and save them to the host machine with `docker build --output`.

2. Building the requirements.txt file is explicitly a part of the
dockerfile for the dev Django image.  From this, we get the benefit of
ensuring that building requirements is done with the same version of
python as the Django environment those requirements will be installed
on.  There is potential for further reducing the duplication of code
between the dev and prod dockerfiles here as well.