Rework directory entry filtering to be based on Django forms
Processing `request.GET` through a Django form gets us some nice
validation features "for free," including rejecting fields that
contain the null character `\x00`.

Another nice thing is we can get `ModelChoiceField`, which does the
querying of the model chosen by the user for us, meaning we can
eliminate most of what the `filters_from_querydict` method was doing.
I've replaced it with a simpler (IMO) mapping between form field names
and filter arguments, which we can iterate over to construct our
search parameters.

Overall, I've tried to improve the tests as well with the addition of
some end-to-end tests of the directory page (rather than tests of the
specific filtering methods).