Don't set `WAGTAILSEARCH_BACKENDS` to `{}`
The settings variable `WAGTAILSEARCH_BACKENDS` is assigned to the
database backend in the base settings file, but in this production
file we are checking to see if ElasticSearch is enabled.  If it
is (not the case, IMO) then we configure the backend, and otherwise
reset the backends setting to `{}` -- I'm not sure how this actually
affects the search feature on the sight, but there's no reason to do
it.  We want the DB search backend in the default case, so let's leave
it that way.