ci: Remove npm audit job requirement that package-lock.json changed
This removes the check on package-lock.json that will cause the npm
audit job to only run if that file was changed.  This was needed to
prevent npm audit from running on PRs when the node packages were not
updated.  But now we are not running it on _any_ PRs, so we can remove
that condition.  It ought to always run when we want it run.