Insert request log middleware earlier in middleware pipeline
This will make it more likely that request and response data is as
correct as possible, such as when the response is intercepted and
changed by another middleware.  For example, the wagtail redirect
middleware will dynamically update 404 reponses to 301 responses, and
if the request logging occurs to late in the pipeline, it will not
have the correct information.