Add structlog's context vars to the production foreign pre-chain
This will make it possible for "foreign" log entries (ones that we do
not generate ourselves, but that are generated by other libraries we
use, such as Django) to include the context variable data we bind to
it in our use of structlog.  One example of this is data we add to
context vars in the `RequestLogMiddleware`.  Data added there will be
lost for events that are logged by Django or other libraries without
this entry in the foreign pre-chain.