Add event codes to github views
I've added the "event_code" key to logging statements here.  I've done
it only for warnings, as (1) we don't really want to aggregate `info`
level events, and (2) `error` level events already have a notification
system that's working well for us.

I have also boosted up some of the warnings to errors, which I did
because their logging messages were capturing the exception
information.  I think this is correct correct to do because,
basically, if we want to log errors, we should log it at the "error"
level.