Removes add, edit and delete options in Scan Result
After moving to ModelViewSet instead of SnippetViewset, I looked into
the wagtail source code and found out a function
`get_common_view_kwargs` which is used to set all the kwargs including
the various URLs like add, edit and delete. This same URL names are
later used by IndexView and the index template as a check to set URL for
editing, or adding the Create button or in the context menu. So I
overrode this function and made all the url names None, which
successfully removes the add, edit and delete options from all places in
the ScanResult admin page